requirements.txt
fastapi
uvicorn
openai
python-multipart
python-dotenv
numpy
google-cloud-storage